+++ CHANGELOG.md
@@ -4,3 +4,4 @@ TinyPNG Image Optimizer Change Log
 Waiting for Release
 -------------------
 - Update "Usage this month" value on Open plugin settings or press Apply
+- Add Background Compress